Issue-347: Moved BaseMapper to core module
